Fix: apparently the tiny LU in bicgstab is not stable enough in float.

+++ benchmarks/physics/benchmark_convection_shell_isovisc_ns_ns.cpp
@@ -43,7 +43,7 @@ using linalg::VectorQ1IsoQ2Q1;
 using linalg::VectorQ1Scalar;
 using linalg::VectorQ1Vec;
 
-using ScalarType = double;
+using ScalarType = float;
 
 struct Parameters
 {
